refactor(frontend-ir): centralize duplicated helpers, kill route-set drift
Both architectural reviews of PR #3128 flagged the frontend_ir tooling's
copy-paste helper sprawl and a live route-set drift hazard. This is
consolidation only — no behavior change (all gate verdicts on the real
Chainer/compressor/static-html artifacts are unchanged; full unit suite
green).

- Add frontend_ir_common.py with the byte-identical helpers that were
  duplicated across the scripts (as_dict ×6, as_list ×7, non_negative_int
  ×4, load_json ×10, write_json ×11) plus load_json_any for the one
  divergent array-accepting loader. Route every script through it.
- primitive_gate.py hand-retyped NATIVE_ROUTES = {"native_cpp", ...} as a
  string literal; it now imports the canonical set from
  frontend_ir_validation so adding a route can't silently desync the
  primitive gate's native-route check. (Reviewers' one must-fix item.)

Net: ~150 lines of duplicated helper bodies removed, single source of
truth for the generic helpers and the native-route set.
