+++ tools/scripts/frontend_ir_codegen_artifact_gate.py
@@ -7,6 +7,7 @@ import argparse
 import json
 import pathlib
 from typing import Any
+from frontend_ir_common import as_dict, as_list, load_json, non_negative_int, write_json
 
 
 PASS_STATUS = "pass"
@@ -17,31 +18,6 @@ NOT_READY_VERDICT = "not_ready"
 CODEGEN_ARTIFACT_SCHEMA = "pulp-frontend-ir-codegen-artifacts-v0"
 GATE_SCHEMA = "pulp-frontend-ir-codegen-artifact-gate-v0"
 NATIVE_CPP_ROUTE = "native_cpp"
-
-
-def load_json(path: pathlib.Path) -> dict[str, Any]:
-    with path.open("r", encoding="utf-8") as f:
-        data = json.load(f)
-    if not isinstance(data, dict):
-        raise ValueError(f"{path} must contain a JSON object")
-    return data
-
-
-def write_json(path: pathlib.Path, data: dict[str, Any]) -> None:
-    path.parent.mkdir(parents=True, exist_ok=True)
-    path.write_text(json.dumps(data, indent=2, sort_keys=True) + "\n", encoding="utf-8")
-
-
-def as_dict(value: Any) -> dict[str, Any]:
-    return value if isinstance(value, dict) else {}
-
-
-def as_list(value: Any) -> list[Any]:
-    return value if isinstance(value, list) else []
-
-
-def non_negative_int(value: Any) -> int:
-    return value if isinstance(value, int) and not isinstance(value, bool) and value >= 0 else 0
 
 
 def check(check_id: str, status: str, message: str, **details: Any) -> dict[str, Any]:
